+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.3.1 - 2021-10-14
+## Added
+* Added a second build stage to invalidate the CloudFront distribution cache.
+
+## Removed
+* Removed the Lambda function that invalidated the CloudFront distribution cache.
+
 ## v0.3.0 - 2021-10-01
 ### Added
 * Added issue templates.